Delete ruff/resolve.rs
