--- crates/ruff/src/resolver.rs
@@ -1,45 +1,0 @@
-//! Discover Python files, and their corresponding [`Settings`], from the
-//! filesystem.
-
-use std::path::PathBuf;
-
-use crate::fs;
-use crate::settings::AllSettings;
-
-/// The configuration information from a `pyproject.toml` file.
-pub struct PyprojectConfig {
-    /// The strategy used to discover the relevant `pyproject.toml` file for
-    /// each Python file.
-    pub strategy: PyprojectDiscoveryStrategy,
-    /// All settings from the `pyproject.toml` file.
-    pub settings: AllSettings,
-    /// Absolute path to the `pyproject.toml` file. This would be `None` when
-    /// either using the default settings or the `--isolated` flag is set.
-    pub path: Option<PathBuf>,
-}
-
-impl PyprojectConfig {
-    pub fn new(
-        strategy: PyprojectDiscoveryStrategy,
-        settings: AllSettings,
-        path: Option<PathBuf>,
-    ) -> Self {
-        Self {
-            strategy,
-            settings,
-            path: path.map(fs::normalize_path),
-        }
-    }
-}
-
-/// The strategy used to discover the relevant `pyproject.toml` file for each
-/// Python file.
-#[derive(Debug, is_macro::Is)]
-pub enum PyprojectDiscoveryStrategy {
-    /// Use a fixed `pyproject.toml` file for all Python files (i.e., one
-    /// provided on the command-line).
-    Fixed,
-    /// Use the closest `pyproject.toml` file in the filesystem hierarchy, or
-    /// the default settings.
-    Hierarchical,
-}
