+++ crates/ruff_cli/src/commands/run_stdin.rs
@@ -4,9 +4,8 @@ use std::path::Path;
 use anyhow::Result;
 
 use ruff::packaging;
-use ruff::resolver::PyprojectConfig;
 use ruff::settings::flags;
-use ruff_workspace::resolver::python_file_at_path;
+use ruff_workspace::resolver::{python_file_at_path, PyprojectConfig};
 
 use crate::args::Overrides;
 use crate::diagnostics::{lint_stdin, Diagnostics};
