+++ crates/ruff_cli/src/commands/show_files.rs
@@ -4,9 +4,8 @@ use std::path::PathBuf;
 use anyhow::Result;
 use itertools::Itertools;
 
-use ruff::resolver::PyprojectConfig;
 use ruff::warn_user_once;
-use ruff_workspace::resolver::python_files_in_path;
+use ruff_workspace::resolver::{python_files_in_path, PyprojectConfig};
 
 use crate::args::Overrides;
