+++ crates/ruff_cli/src/resolve.rs
@@ -4,10 +4,12 @@ use anyhow::Result;
 use log::debug;
 use path_absolutize::path_dedot;
 
-use ruff::resolver::{PyprojectConfig, PyprojectDiscoveryStrategy};
 use ruff::settings::configuration::Configuration;
 use ruff::settings::{pyproject, AllSettings};
-use ruff_workspace::resolver::{resolve_settings_with_processor, ConfigProcessor, Relativity};
+use ruff_workspace::resolver::{
+    resolve_settings_with_processor, ConfigProcessor, PyprojectConfig, PyprojectDiscoveryStrategy,
+    Relativity,
+};
 
 use crate::args::Overrides;
